backend/requirements.txt
#Version django
Django>=4.0.4


#mysqlclient
mysqlclient>=2.1.1

#Rest
djangorestframework>=3.14.0

#Pillow
Pillow>=9.2.0


#django-cors-headers
django-cors-headers>=3.13.0


#whitenoise
whitenoise>=6.2.0

#gunicorn

gunicorn>=20.1.0

#tzdata

tzdata>= 2022.7


#decouple
python-decouple>= 3.7

#DOTENV

python-dotenv>= 1.0.0



#simple catcha
django-simple-captcha>= 0.5.17

#requests
requests>= 2.28.2

